# artifacts/api-server/src/routes/messages.ts
import { Router, type IRouter } from "express";
import { and, asc, eq, inArray } from "drizzle-orm";
import {
  db,
  messagesTable,
  messageReactionsTable,
  usersTable,
  groupsTable,
} from "@workspace/db";
import {
  ListMessagesResponseItem,
  SendMessageBody,
  SendMessageResponse,
  EditMessageBody,
  EditMessageResponse,
  DeleteMessageResponse,
  ToggleMessageReactionBody,
  ToggleMessageReactionResponse,
} from "@workspace/api-zod";
import { requireAuth } from "../middlewares/requireAuth";
import { messageSendRateLimit } from "../middlewares/rateLimit";
import {
  parseGroupId,
  isGroupMember,
  getGroupMemberIds,
} from "../lib/groupAccess";
import { broadcastToGroup, isUserOnline } from "../ws/hub";
import { sendPushToUser } from "../lib/push";
import { toIso, toIsoOrNull } from "../lib/serialize";

const router: IRouter = Router();

router.use(requireAuth);

/**
 * Aggregates raw (messageId, userId, emoji) reaction rows into the
 * `{ emoji, userIds }[]` shape the API returns — one entry per distinct
 * emoji on a message, with everyone who reacted with it.
 */
async function getReactionsByMessageId(
  messageIds: number[],
): Promise<Map<number, { emoji: string; userIds: string[] }[]>> {
  if (messageIds.length === 0) return new Map();

  const rows = await db
    .select({
      messageId: messageReactionsTable.messageId,
      userId: messageReactionsTable.userId,
      emoji: messageReactionsTable.emoji,
    })
    .from(messageReactionsTable)
    .where(inArray(messageReactionsTable.messageId, messageIds));

  const byMessage = new Map<number, Map<string, string[]>>();
  for (const row of rows) {
    const byEmoji = byMessage.get(row.messageId) ?? new Map<string, string[]>();
    byEmoji.set(row.emoji, [...(byEmoji.get(row.emoji) ?? []), row.userId]);
    byMessage.set(row.messageId, byEmoji);
  }

  const result = new Map<number, { emoji: string; userIds: string[] }[]>();
  for (const [messageId, byEmoji] of byMessage) {
    result.set(
      messageId,
      Array.from(byEmoji.entries()).map(([emoji, userIds]) => ({
        emoji,
        userIds,
      })),
    );
  }
  return result;
}

/**
 * Batch-fetches a denormalized reply preview for every distinct replyToId
 * among the given messages, keyed by that replyToId. One extra query
 * regardless of how many messages are replies, same batching approach as
 * getReactionsByMessageId above.
 */
async function getReplyPreviewsByReplyToId(replyToIds: number[]): Promise<
  Map<
    number,
    {
      id: string;
      senderId: string;
      senderName: string;
      content: string;
      type: string;
      fileName: string | null;
      deletedAt: string | null;
    }
  >
> {
  const uniqueIds = [...new Set(replyToIds)];
  if (uniqueIds.length === 0) return new Map();

  const rows = await db
    .select({
      id: messagesTable.id,
      senderId: messagesTable.senderId,
      senderName: usersTable.name,
      content: messagesTable.content,
      type: messagesTable.type,
      fileName: messagesTable.fileName,
      deletedAt: messagesTable.deletedAt,
    })
    .from(messagesTable)
    .innerJoin(usersTable, eq(messagesTable.senderId, usersTable.id))
    .where(inArray(messagesTable.id, uniqueIds));

  const result = new Map<
    number,
    {
      id: string;
      senderId: string;
      senderName: string;
      content: string;
      type: string;
      fileName: string | null;
      deletedAt: string | null;
    }
  >();
  for (const row of rows) {
    result.set(row.id, {
      id: String(row.id),
      senderId: row.senderId,
      senderName: row.senderName,
      content: row.deletedAt ? "" : row.content,
      type: row.type,
      fileName: row.deletedAt ? null : row.fileName,
      deletedAt: toIsoOrNull(row.deletedAt),
    });
  }
  return result;
}

function parsePaginationQuery(query: Record<string, unknown>): {
  limit: number;
  offset: number;
} {
  const rawLimit = Number(query.limit);
  const limit =
    Number.isFinite(rawLimit) && rawLimit > 0 && rawLimit <= 100
      ? Math.floor(rawLimit)
      : 50;

  const rawOffset = Number(query.offset);
  const offset =
    Number.isFinite(rawOffset) && rawOffset >= 0 ? Math.floor(rawOffset) : 0;

  return { limit, offset };
}

router.get("/groups/:groupId/messages", async (req, res): Promise<void> => {
  const groupId = parseGroupId(req.params.groupId);
  if (groupId === null) {
    res.status(404).json({ error: "Group not found" });
    return;
  }

  const member = await isGroupMember(groupId, req.userId!);
  if (!member) {
    res.status(404).json({ error: "Group not found" });
    return;
  }

  const { limit, offset } = parsePaginationQuery(req.query);

  const rows = await db
    .select({
      id: messagesTable.id,
      groupId: messagesTable.groupId,
      senderId: messagesTable.senderId,
      senderName: usersTable.name,
      senderAvatarUrl: usersTable.avatarUrl,
      content: messagesTable.content,
      type: messagesTable.type,
      fileName: messagesTable.fileName,
      mimeType: messagesTable.mimeType,
      fileSize: messagesTable.fileSize,
      durationSeconds: messagesTable.durationSeconds,
      replyToId: messagesTable.replyToId,
      mentionedUserIds: messagesTable.mentionedUserIds,
      createdAt: messagesTable.createdAt,
      editedAt: messagesTable.editedAt,
      deletedAt: messagesTable.deletedAt,
    })
    .from(messagesTable)
    .innerJoin(usersTable, eq(messagesTable.senderId, usersTable.id))
    .where(eq(messagesTable.groupId, groupId))
    .orderBy(asc(messagesTable.createdAt))
    .limit(limit)
    .offset(offset);

  const reactionsByMessageId = await getReactionsByMessageId(
    rows.map((r) => r.id),
  );
  const replyPreviewsByReplyToId = await getReplyPreviewsByReplyToId(
    rows.flatMap((r) => (r.replyToId ? [r.replyToId] : [])),
  );

  res.json(
    rows.map((row) =>
      ListMessagesResponseItem.parse({
        ...row,
        id: String(row.id),
        groupId: String(row.groupId),
        replyToId: row.replyToId ? String(row.replyToId) : null,
        replyTo: row.replyToId
          ? (replyPreviewsByReplyToId.get(row.replyToId) ?? null)
          : null,
        createdAt: toIso(row.createdAt),
        editedAt: toIsoOrNull(row.editedAt),
        deletedAt: toIsoOrNull(row.deletedAt),
        reactions: reactionsByMessageId.get(row.id) ?? [],
      }),
    ),
  );
});

router.post(
  "/groups/:groupId/messages",
  messageSendRateLimit,
  async (req, res): Promise<void> => {
    const groupId = parseGroupId(req.params.groupId);
    if (groupId === null) {
      res.status(404).json({ error: "Group not found" });
      return;
    }

    const userId = req.userId!;
    const member = await isGroupMember(groupId, userId);
    if (!member) {
      res.status(404).json({ error: "Group not found" });
      return;
    }

    const parsed = SendMessageBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    // A replyToId must point at a real message in this same group — a
    // dangling/foreign replyToId is silently dropped rather than trusted,
    // same defensive posture as the mention validation below.
    let replyToId: number | null = null;
    if (parsed.data.replyToId) {
      const candidate = Number.parseInt(parsed.data.replyToId, 10);
      if (Number.isFinite(candidate)) {
        const [target] = await db
          .select({ id: messagesTable.id })
          .from(messagesTable)
          .where(
            and(
              eq(messagesTable.id, candidate),
              eq(messagesTable.groupId, groupId),
            ),
          );
        if (target) replyToId = target.id;
      }
    }

    // Mentions are plaintext metadata sent alongside the encrypted content
    // (see messagesTable.mentionedUserIds) — only accept IDs that are
    // actually current members of this group.
    const memberIds = await getGroupMemberIds(groupId);
    const mentionedUserIds = (parsed.data.mentionedUserIds ?? []).filter((id) =>
      memberIds.has(id),
    );

    const [message] = await db
      .insert(messagesTable)
      .values({
        groupId,
        senderId: userId,
        content: parsed.data.content,
        type: parsed.data.type ?? "text",
        fileName: parsed.data.fileName ?? null,
        mimeType: parsed.data.mimeType ?? null,
        fileSize: parsed.data.fileSize ?? null,
        durationSeconds: parsed.data.durationSeconds ?? null,
        replyToId,
        mentionedUserIds,
      })
      .returning();

    const [sender] = await db
      .select({ name: usersTable.name, avatarUrl: usersTable.avatarUrl })
      .from(usersTable)
      .where(eq(usersTable.id, userId));

    const replyTo = message.replyToId
      ? ((await getReplyPreviewsByReplyToId([message.replyToId])).get(
          message.replyToId,
        ) ?? null)
      : null;

    const payload = SendMessageResponse.parse({
      id: String(message.id),
      groupId: String(message.groupId),
      senderId: message.senderId,
      senderName: sender?.name ?? "Family Member",
      senderAvatarUrl: sender?.avatarUrl ?? null,
      content: message.content,
      type: message.type,
      fileName: message.fileName,
      mimeType: message.mimeType,
      fileSize: message.fileSize,
      durationSeconds: message.durationSeconds,
      replyToId: message.replyToId ? String(message.replyToId) : null,
      replyTo,
      mentionedUserIds: message.mentionedUserIds,
      createdAt: toIso(message.createdAt),
      editedAt: null,
      deletedAt: null,
      reactions: [],
    });

    broadcastToGroup(groupId, { type: "message", message: payload });

    // Only to members who aren't currently online (any tab/device) — if
    // they are, the WS message above already reached them live. Never
    // include message content, same reasoning as the DM version of this.
    const [group] = await db
      .select({ name: groupsTable.name })
      .from(groupsTable)
      .where(eq(groupsTable.id, groupId));
    for (const memberId of memberIds) {
      if (memberId === userId || isUserOnline(memberId)) continue;
      void sendPushToUser(memberId, {
        title: group?.name ?? "Family Group",
        body: `${sender?.name ?? "Someone"} sent a new message`,
        url: `/app/groups/${groupId}`,
      });
    }

    res.status(201).json(payload);
  },
);

router.patch(
  "/groups/:groupId/messages/:messageId",
  async (req, res): Promise<void> => {
    const groupId = parseGroupId(req.params.groupId);
    const messageId = parseGroupId(req.params.messageId);
    if (groupId === null || messageId === null) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const userId = req.userId!;
    const member = await isGroupMember(groupId, userId);
    if (!member) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const [existing] = await db
      .select()
      .from(messagesTable)
      .where(
        and(
          eq(messagesTable.id, messageId),
          eq(messagesTable.groupId, groupId),
        ),
      );
    if (!existing) {
      res.status(404).json({ error: "Message not found" });
      return;
    }
    if (existing.senderId !== userId) {
      res.status(403).json({ error: "You can only edit your own messages" });
      return;
    }
    if (existing.type !== "text" || existing.deletedAt) {
      res.status(403).json({ error: "This message can't be edited" });
      return;
    }

    const parsed = EditMessageBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    const [updated] = await db
      .update(messagesTable)
      .set({ content: parsed.data.content, editedAt: new Date() })
      .where(eq(messagesTable.id, messageId))
      .returning();

    const [sender] = await db
      .select({ name: usersTable.name, avatarUrl: usersTable.avatarUrl })
      .from(usersTable)
      .where(eq(usersTable.id, userId));

    const reactions =
      (await getReactionsByMessageId([messageId])).get(messageId) ?? [];
    const replyTo = updated.replyToId
      ? ((await getReplyPreviewsByReplyToId([updated.replyToId])).get(
          updated.replyToId,
        ) ?? null)
      : null;

    const payload = EditMessageResponse.parse({
      id: String(updated.id),
      groupId: String(updated.groupId),
      senderId: updated.senderId,
      senderName: sender?.name ?? "Family Member",
      senderAvatarUrl: sender?.avatarUrl ?? null,
      content: updated.content,
      type: updated.type,
      fileName: updated.fileName,
      mimeType: updated.mimeType,
      fileSize: updated.fileSize,
      durationSeconds: updated.durationSeconds,
      replyToId: updated.replyToId ? String(updated.replyToId) : null,
      replyTo,
      mentionedUserIds: updated.mentionedUserIds,
      createdAt: toIso(updated.createdAt),
      editedAt: toIsoOrNull(updated.editedAt),
      deletedAt: toIsoOrNull(updated.deletedAt),
      reactions,
    });

    broadcastToGroup(groupId, { type: "message-updated", message: payload });

    res.json(payload);
  },
);

router.delete(
  "/groups/:groupId/messages/:messageId",
  async (req, res): Promise<void> => {
    const groupId = parseGroupId(req.params.groupId);
    const messageId = parseGroupId(req.params.messageId);
    if (groupId === null || messageId === null) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const userId = req.userId!;
    const member = await isGroupMember(groupId, userId);
    if (!member) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const [existing] = await db
      .select()
      .from(messagesTable)
      .where(
        and(
          eq(messagesTable.id, messageId),
          eq(messagesTable.groupId, groupId),
        ),
      );
    if (!existing) {
      res.status(404).json({ error: "Message not found" });
      return;
    }
    if (existing.senderId !== userId) {
      res.status(403).json({ error: "You can only delete your own messages" });
      return;
    }

    const [updated] = await db
      .update(messagesTable)
      .set({
        content: "",
        fileName: null,
        mimeType: null,
        fileSize: null,
        deletedAt: new Date(),
      })
      .where(eq(messagesTable.id, messageId))
      .returning();

    const [sender] = await db
      .select({ name: usersTable.name, avatarUrl: usersTable.avatarUrl })
      .from(usersTable)
      .where(eq(usersTable.id, userId));

    const reactions =
      (await getReactionsByMessageId([messageId])).get(messageId) ?? [];
    const replyTo = updated.replyToId
      ? ((await getReplyPreviewsByReplyToId([updated.replyToId])).get(
          updated.replyToId,
        ) ?? null)
      : null;

    const payload = DeleteMessageResponse.parse({
      id: String(updated.id),
      groupId: String(updated.groupId),
      senderId: updated.senderId,
      senderName: sender?.name ?? "Family Member",
      senderAvatarUrl: sender?.avatarUrl ?? null,
      content: updated.content,
      type: updated.type,
      fileName: updated.fileName,
      mimeType: updated.mimeType,
      fileSize: updated.fileSize,
      durationSeconds: updated.durationSeconds,
      replyToId: updated.replyToId ? String(updated.replyToId) : null,
      replyTo,
      mentionedUserIds: updated.mentionedUserIds,
      createdAt: toIso(updated.createdAt),
      editedAt: toIsoOrNull(updated.editedAt),
      deletedAt: toIsoOrNull(updated.deletedAt),
      reactions,
    });

    broadcastToGroup(groupId, { type: "message-updated", message: payload });

    res.json(payload);
  },
);

router.put(
  "/groups/:groupId/messages/:messageId/reactions",
  async (req, res): Promise<void> => {
    const groupId = parseGroupId(req.params.groupId);
    const messageId = parseGroupId(req.params.messageId);
    if (groupId === null || messageId === null) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const userId = req.userId!;
    const member = await isGroupMember(groupId, userId);
    if (!member) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const [existingMessage] = await db
      .select({ id: messagesTable.id })
      .from(messagesTable)
      .where(
        and(
          eq(messagesTable.id, messageId),
          eq(messagesTable.groupId, groupId),
        ),
      );
    if (!existingMessage) {
      res.status(404).json({ error: "Message not found" });
      return;
    }

    const parsed = ToggleMessageReactionBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }
    const { emoji } = parsed.data;

    const [existingReaction] = await db
      .select({ id: messageReactionsTable.id })
      .from(messageReactionsTable)
      .where(
        and(
          eq(messageReactionsTable.messageId, messageId),
          eq(messageReactionsTable.userId, userId),
          eq(messageReactionsTable.emoji, emoji),
        ),
      );

    if (existingReaction) {
      await db
        .delete(messageReactionsTable)
        .where(eq(messageReactionsTable.id, existingReaction.id));
    } else {
      await db
        .insert(messageReactionsTable)
        .values({ messageId, userId, emoji });
    }

    const reactions =
      (await getReactionsByMessageId([messageId])).get(messageId) ?? [];

    // Reuse the existing "message-updated" WS event so every client's
    // already-wired listener picks up the new reaction list without a new
    // event type — same idea as an edit, just a different field changing.
    const [full] = await db
      .select()
      .from(messagesTable)
      .where(eq(messagesTable.id, messageId));
    const [sender] = await db
      .select({ name: usersTable.name, avatarUrl: usersTable.avatarUrl })
      .from(usersTable)
      .where(eq(usersTable.id, full.senderId));

    const replyTo = full.replyToId
      ? ((await getReplyPreviewsByReplyToId([full.replyToId])).get(
          full.replyToId,
        ) ?? null)
      : null;

    broadcastToGroup(groupId, {
      type: "message-updated",
      message: SendMessageResponse.parse({
        id: String(full.id),
        groupId: String(full.groupId),
        senderId: full.senderId,
        senderName: sender?.name ?? "Family Member",
        senderAvatarUrl: sender?.avatarUrl ?? null,
        content: full.content,
        type: full.type,
        fileName: full.fileName,
        mimeType: full.mimeType,
        fileSize: full.fileSize,
        durationSeconds: full.durationSeconds,
        replyToId: full.replyToId ? String(full.replyToId) : null,
        replyTo,
        mentionedUserIds: full.mentionedUserIds,
        createdAt: toIso(full.createdAt),
        editedAt: toIsoOrNull(full.editedAt),
        deletedAt: toIsoOrNull(full.deletedAt),
        reactions,
      }),
    });

    res.json(ToggleMessageReactionResponse.parse(reactions));
  },
);

export default router;