# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "0.2.0"
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a group
      description: >
        Returns messages oldest-first. Use limit/offset query params for pagination.
        Defaults to 50 messages per page, maximum 100.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of messages, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Rate limited — sending messages too quickly
  /groups/{groupId}/messages/{messageId}:
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a text message you sent
      description: >
        Only the original sender can edit their own message, and only
        text messages can be edited (not file attachments).
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageEditInput"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Not the sender, or not a text message
        "404":
          description: Not found
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message you sent
      description: >
        Soft-deletes the message: content and any attachment are cleared,
        but the message row remains (shown as "This message was deleted").
        Only the original sender can delete their own message.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted message (soft-deleted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Not the sender
        "404":
          description: Not found
  /groups/{groupId}/messages/{messageId}/reactions:
    put:
      operationId: toggleMessageReaction
      tags: [messages]
      summary: Toggle your reaction on a message
      description: >
        If you've already reacted with this emoji, removes it; otherwise
        adds it. Any current group member may react, including to their own
        messages. Returns the message's full updated reaction list.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleReactionInput"
      responses:
        "200":
          description: Updated reactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageReaction"
        "404":
          description: Not found
...
components:
  schemas:
...
    ToggleReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
          minLength: 1
          maxLength: 8
...
    MessageReaction:
      type: object
      required: [emoji, userIds]
      properties:
        emoji:
          type: string
        userIds:
          type: array
          items:
            type: string
    MessageReplyPreview:
      type: object
      description: >
        A denormalized snapshot of the message being replied to, so clients
        can render a quoted snippet without needing that message to already
        be loaded on the page. `content` is still E2E-encrypted ciphertext
        (the server never sees plaintext), decrypted client-side same as
        any other message.
      required: [id, senderId, senderName, content, type, deletedAt]
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, file, voice]
        fileName:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
          description: If set, the quoted message was deleted; show a placeholder instead of the (now-cleared) content.
    Message:
      type: object
      required:
        [
          id,
          groupId,
          senderId,
          senderName,
          content,
          type,
          createdAt,
          reactions,
          mentionedUserIds,
        ]
      properties:
        id:
          type: string
        groupId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        type:
          type: string
          enum: [text, file, voice, system, call]
          default: text
        fileName:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        durationSeconds:
          type: ["integer", "null"]
          description: Playback duration in seconds, for voice messages.
        replyToId:
          type: ["string", "null"]
        replyTo:
          oneOf:
            - $ref: "#/components/schemas/MessageReplyPreview"
            - type: "null"
        mentionedUserIds:
          type: array
          items:
            type: string
          description: Group member user IDs tagged with @mentions in this message.
        createdAt:
          type: string
        editedAt:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/MessageReaction"
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum: [text, file, voice]
          default: text
        fileName:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        durationSeconds:
          type: ["integer", "null"]
        replyToId:
          type: ["string", "null"]
        mentionedUserIds:
          type: array
          items:
            type: string
    MessageEditInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...